Clean up task call stack size handling.
[Reviewed by Sung.]

Currently we handle task call stack sizes in a somewhat fractured way in
the runtime.  There is a function common shared by all tasking layers
that collects from the environment any call stack size specified by the
user, and then each tasking layer implementation uses this value (if
any) and its own default to select an actual call stack size.

This commit changes that somewhat, moving the specification of a default
call stack size into the common code of the runtime.  Tasking layers are
expected to implement task call stacks that are no smaller than the size
given by this common code.

While here, I also changed the type of the task call stack size to
size_t, and did some other cleanups.

The specific file changes are as follows:

runtime/include/chpl-tasks.h
runtime/src/chpl-tasks.c
  Rename chpl_task_getenvCallStackSize() as chpl_task_getMinCallStackSize()
  and change its return type from uint64_t to size_t.  Change it to return
  the minimum call stack size rather than the size specified by the user in
  the environment, if any.  Make the default call stack size 8 MiB.

runtime/include/chpl-threads.h
  Change the type of chpl_threads_getCallStackSize() from uint64_t to
  size_t.

runtime/src/tasks/fifo/tasks-fifo.c
  Change the type of chpl_task_getCallStackSize() from uint64_t to
  size_t.

runtime/src/tasks/massivethreads/tasks-massivethreads.c
  Remove logic that handles the default call stack size, since this is
  now done in the common code.  Change chpl_task_getCallStackSize() from
  uint64_t to size_t.

runtime/src/threads/pthreads/threads-pthreads.c
  Rework the call stack size logic, since the common code now provides a
  default value.  Change chpl_thread_getCallStackSize() from uint64_t to
  size_t.

third-party/qthread/qthread-1.10/src/interfaces/chapel/tasks-qthreads.c
  Rework the call stack size logic, since the common code now provides a
  default value.  Change chpl_task_getCallStackSize() from uint64_t to
  size_t.

modules/internal/localeModels/numa/LocaleModel.chpl
modules/internal/localeModels/flat/LocaleModel.chpl
  Change here.callStackSize and chpl_task_getCallStackSize() from
  uint64_t to size_t.

modules/internal/ChapelThreads.chpl
  Remove callStackSize and extern proc chpl_task_getenvCallStackSize(),
  which is used to initialize it.  This variable was not being used and
  is not documented anywhere.  Plus, since we've said that the call
  stack size is at least potentially different on each locale, it's hard
  to see how a global variable containing its value could be useful.

test/execflags/bradc/callStackSize.prediff
test/execflags/bradc/callStackSize.good (deleted, moved to .good.in)
test/execflags/bradc/callStackSize.good.in (new)
test/execflags/bradc/callStackSize.comm-none.good (deleted, moved to .good.in)
test/execflags/bradc/callStackSize.comm-none.good.in (new)
test/execflags/bradc/callStackSize.cleanfiles
  The output for this test varies depending on which tasking layer it's
  built with, because some tasking layers (fifo, for example) round up
  the call stack size to the next system page boundary, while others do
  not.  To handle this, add a .prediff script to create the appropriate
  .good for the specific test environment.  Move the existing .good
  files to new .good.in files containing the output that would match the
  stack size specified in the .execenv file.  The new .prediff processes
  these as necessary to match the specific case.  The new .cleanfiles
  file removes the dynamically created .good file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22771 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/include/chpl-threads.h
@@ -85,7 +85,7 @@ uint32_t chpl_thread_getNumThreads(void);
 //
 // Get the current thread stack size.
 //
-uint64_t chpl_thread_getCallStackSize(void);
+size_t chpl_thread_getCallStackSize(void);
 
 //
 // Mutexes, and operations upon them.
